fix: libnotify freeze when calling dbus notify
The calls to adjust geometry were executed too often, results in
flicking bubble and the caller (libnotify) to freeze until a dbus
timeout, which is often seen in tg-desktop. The adjustment should be
done when handling corresponding signals, if I understand it correctly.

Removing the two calls fixes the freeze, and makes the bubble much more
smooth (show & hide animation). I have tried to place dde-dock in the
four conners, and also make dde-control-center to show/hide, turns out
the position of the bubble remains correct.

Change-Id: If786b9587a587bc7c1e06d5920c9586a34b256fd